+++ src/pages/MentorPage/MentorPage.css
@@ -7,6 +7,10 @@
     }
 }
 
+#mentor {
+    background-color: #FFFDFA;
+}
+
 .mentor-image {
     object-fit: cover;
     width:150px;
