+++ src/pages/MentorPage/MentorPage.js
@@ -1,87 +1,21 @@
 import React from "react";
-import Image from 'react-bootstrap/Image';
-import m1 from '../../assets/images/MentorPage/KS.jpg';
-import m2 from '../../assets/images/MentorPage/DS.PNG';
-import m3 from '../../assets/images/MentorPage/Karen.png';
-import m4 from '../../assets/images/MentorPage/DC.jpg';
-import m5 from '../../assets/images/MentorPage/NS.png';
-import Container from "react-bootstrap/Container";
-import Row from "react-bootstrap/Row";
-import Col from "react-bootstrap/Col";
-//import Card from 'react-bootstrap/Card';
+import MobileMentorPage from './components/MobileMentorPage';
+import DesktopMentorPage from './components/DesktopMentorPage';
 import "./MentorPage.css";
-const mentors=[
-{
-    imageSrc:m1,
-    name:"Kristina Stephens",
-    role:"Student Health And Wellness"
-},
-{
-    imageSrc:m2,
-    name:"Daniel Sandberg",
-    role:"Student Health And Wellness"
-},
-{
-    imageSrc:m3,
-    name:"Karen McComb",
-    role:"Director of Health Promotion and Community Impact for Student Health And Wellness at UConn"
-},
 
-{
-    imageSrc:m4,
-    name:"David Coles",
-    role:"Entrepreneur, Wealth Advisor, and CEO of Human Wealth"
-
-},
-{
-    imageSrc:m5,
-    name:"Nick Satta",
-    role:"Lab Manager, OPIM Innovate"
+function isMobile() {
+    if (window.innerWidth <= 576) {
+        return true;
+    }
+    else {
+        return false;
+    }
 }
-]
+
+
 class MentorPage extends React.Component {
     render() {
-        return(
-            <div className="pt-5">
-                <Container>
-                    <Row className="mt-5">
-                        <h1 className="aboutHeader">Meet The Mentors!</h1>
-                        <Col sm={4}>
-                            <Col><Image alt="KS" className="mentor-image" src={m1} /></Col>
-                            <Col><h1 className="name-text">Kristina Stephens</h1></Col>
-                            <Col><p className="role-text">Director of Mental Health at University of Connecticut</p></Col>
-                        </Col>
-                        <Col sm={4}>
-                            <Col><Image alt="DS" className="mentor-image" src={m2} /></Col>
-                            <Col><h1 className="name-text">Daniel Sandberg</h1></Col>
-                            <Col><p className="role-text">Senior Director, Quantitative Research at S&P Global Market Intelligence</p></Col>
-                        </Col>
-                        <Col sm={4}>
-                            <Col><Image alt="KS" className="mentor-image" src={m3} /></Col>
-                            <Col><h1 className="name-text">Karen McComb</h1></Col>
-                            <Col><p className="role-text">Director of Health Promotion and Community Impact for Student Health And Wellness at UConn</p></Col>
-                        </Col>
-                    </Row>
-                    <Row className="mt-5">
-                    <Col sm={4}>
-                            <Col><Image alt="KS" className="mentor-image" src={m4} /></Col>
-                            <Col><h1 className="name-text">David Coles</h1></Col>
-                            <Col><p className="role-text">Entrepreneur, Wealth Advisor, and CEO of Human Wealth</p></Col>
-                        </Col>
-                        <Col sm={4}>
-                            <Col><Image alt="KS" className="mentor-image" src={m5} /></Col>
-                            <Col><h1 className="name-text">Nick Satta</h1></Col>
-                            <Col><p className="role-text">Lab Manager, OPIM Innovate</p></Col>
-                        </Col>
-                        
-                    </Row>
-                    
-                </Container>
-                <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 1440 320">
-                <path fill="#815AE0" fill-opacity="1" d="M0,128L60,144C120,160,240,192,360,186.7C480,181,600,139,720,117.3C840,96,960,96,1080,101.3C1200,107,1320,117,1380,122.7L1440,128L1440,320L1380,320C1320,320,1200,320,1080,320C960,320,840,320,720,320C600,320,480,320,360,320C240,320,120,320,60,320L0,320Z"></path>
-            </svg>
-            </div>
-        ) 
+        return isMobile() ? ( <MobileMentorPage /> ) : ( <DesktopMentorPage /> )
     }
 }
 export default MentorPage;
