+++ src/pages/SponsorsPage/SponsorsPage.css
@@ -1,5 +1,3 @@
-
-
 .platinumGear {
     max-height: 15rem;
     display: block;
